requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorzero==2.0
configparser==7.1.0
idna==3.10
lgpio==0.2.2.0
pyserial==3.5
requests==2.32.3
rpi-lgpio==0.6
urllib3==2.2.3
